requirements.txt
anthropic>=0.40.0
openpyxl>=3.1.0
PyMuPDF>=1.24.0
mkdocs-material>=9.5.0
pyyaml>=6.0
